Completato libreria 100% test passati

+++ Esercizio 1/Exercise1/exercise1/queue/vec/queuevec.hpp
@@ -20,9 +20,9 @@ class QueueVec : virtual public Queue<Data>,virtual public Vector<Data>{
 
 private:
 
-static const int DEFAULT_SIZE = 4;
-static const int INCREASE_FACTOR = 2;
-static const int DECREASE_FACTOR = 2;
+static const unsigned long DEFAULT_SIZE = 4;
+static const unsigned long INCREASE_FACTOR = 2;
+static const unsigned long DECREASE_FACTOR = 2;
   
 
 protected:
